fix(analytics): repair marketing dashboard queries (activity 500 + wrong source)
- activity: fix CardinalityViolation (avg over a GROUP BY'd scalar subquery →
  FROM-clause subquery); cast avg to Numeric for round(); widen unused DTO ints
  → float. DAU/MAU/total/active/stickiness now compute.
- payments/info + top_clients: re-point from the never-emitted purchase_success
  event stream to the real  table (status/created_at), методы по pg_payment_method.
- retention: add the per-month  field the chart expects.
Note: top-level registrations/retention stay 0 until a user_registered event is emitted (no source).

+++ src/analytics/dtos/activity.py
@@ -32,8 +32,8 @@ class AUDTO(BaseModel):
 class ActivityDTO(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
-    avg_time_per_session: int = 0
-    avg_session_per_day: int = 0
+    avg_time_per_session: float = 0
+    avg_session_per_day: float = 0
 
     dau_mau_ratio: float = 0.0
     mau_dau_ratio: float = 0.0
